git-svn-id: https://speak-lib.googlecode.com/svn/trunk@13 63a8470e-8155-0410-a357-b11873b86f21

+++ TrueOrFalse.Core/Domain/Question/Answer/Answer.cs
@@ -7,6 +7,6 @@ namespace SpeakFriend.TrueOrFalse
 {
     public class Answer
     {
-        public double Value;
+        public int Value;
     }
 }
